QA fixes
/login endpoint returns a valid token for any POST, wether credentials
are supplied or not.

FIX = Just checking for empty body. Note that extensive validation is
not implemented at this time

/users/register works the first time but the user that is created has
an id of "58381630-00ff-11e6-ac9c-f74253bf2c0a” and no other fields

FIX = remove "updatedBy": "admin", "createdBy": "admin"

The created user w/that id cannot be deleted (404 response w/“No id
found…” message).

FIX = from this point on all ‘values’ must be strings. The last version
we stuck a number here .

Subsequent register posts fail w/an “Email already taken” error whether
a unique email is supplied or not.

FIX = could not reproduce this. What may have been happening is that
you can NOT supply and email and get passed. Again this is simply
because validation is not implemented on all requests yet.

There are no methods for Updating a user’s information, one of the
Requirements of the challenge.  We were looking for a PUT or PATCH
method for an existing user.

FIX = quickly adding a PUT endpoint to allow changing users name filed.
Only name currently.